myapp/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
requests==2.32.3
websocket-client==1.6.3
Werkzeug==2.3.7
Pillow==10.0.1
python-socketio==5.9.0
python-engineio==4.7.1